+++ src/core/Match.js
@@ -3,11 +3,17 @@ module.exports = Match;
 var config = require('../config.js');
 var util = require('../util.js');
 
-function Match(fixture) {
-    this.data = fixture;
-    this.events = [];
-    this.currentEvent = -1;
-    this.$el = $('#' + util.matchId(this.data.dbid));
+/**
+ * A match holds a single fixture between two teams
+ * @param {object} matchData - Raw data from the json database
+ * @constructor
+ */
+function Match(matchData) {
+    /**
+     * Match data from the json database
+     * @type {Object}
+     */
+    this.data = matchData;
 
     //Sort the matchevents by time
     this.data.matchevents.sort(function(a,b) {
@@ -22,6 +28,11 @@ function Match(fixture) {
     });
 }
 
+/**
+ * Get all goals from the match in chronological order
+ * @returns {Array.<object>} - An array of all match events with goals
+ * //TODO generalize for any available stat
+ */
 Match.prototype.getGoals = function() {
     var goals = [];