+++ src/core/ReplayEngine.js
@@ -6,18 +6,65 @@ var util = require('../util.js');
 
 var Competition = require('./Competition.js');
 
-
+/**
+ * Main module for the League Replay
+ * @param {number} compId - Id of the competition to load
+ * @constructor
+ */
 function ReplayEngine(compId) {
+    /**
+     * Self reference
+     * @type {ReplayEngine}
+     */
     var _this = this;
-
+    /**
+     * Holds all data for the competition
+     * @type {Competition}
+     */
     this.competition = new Competition(compId);
+    /**
+     * Holds events that are ready for execution
+     * @type {Array.<object>}
+     */
+    this.events = [];
+    /**
+     * Index of the current event
+     * @type {number}
+     */
+    this.eventCounter = -1;
+    /**
+     * Id of the match to simulate up until
+     * @type {number}
+     */
+    this.simulateTo = -1;
+    /**
+     * Match currently animating
+     * @type {Match}
+     */
+    this.currentMatch = null;
+    /**
+     * Reference to the element of the match currently animating
+     * @type {jQuery}
+     */
+    this.$currentMatch = null;
+    /**
+     * Used to check if video is playing
+     * TODO check if this is actually used
+     * @type {boolean}
+     */
+    this.videoIsPlaying = false;
+    /**
+     * Reference to the fixtures element
+     * @type {jQuery}
+     */
+    this.$fixtures = $('#fixtures');
+
+    //Render the view when the competition data has loaded
     this.competition.on('ready', function() {
-        this.render();
-    });
-    this.competition.on('finish', function() {
-        console.log("Competition has finished animating");
+        _this.competition.render();
     });
 
+    //Add teams to follow, when clicking team names
     $(document).on('click', '.' + config.class.FOLLOW_TEAM, function(e) {
         e.preventDefault();
         if($(this).hasClass(config.class.FOLLOWED)) {
@@ -27,6 +74,8 @@ function ReplayEngine(compId) {
         }
     });
 
+    //Select a match to jump to, when clicking the match selecotr
+    //TODO should be able to do it in reverse - jump to any point in the season
     $(document).one('click', '.' + config.class.MATCH_SELECT, function(e) {
         e.preventDefault();
         selectMatch(this);
@@ -47,10 +96,6 @@ function ReplayEngine(compId) {
             $('.' + config.class.MATCH_SELECTED).removeClass(config.class.MATCH_SELECTED);
             _this.simulateTo = $(el).parent().attr('id');
             $(el).addClass(config.class.MATCH_SELECTED);
-
-            setTimeout(function() {
-                $(el).removeClass(config.class.MATCH_SELECTED);
-            }, 3000); //TODO move to settings file
         }
     });
 
@@ -75,37 +120,35 @@ function ReplayEngine(compId) {
             }
         });
 
+        //Starts the event loop
         _this.processMatch(_this.competition.nextMatch());
         _this.nextEvent();
     });
-
-    this.eventCounter = -1;
-    this.events = [];
-
-    this.simulateTo = -1;
-
-    this.currentMatch = null;
-    this.$currentMatch = null;
-
-    this.videoIsPlaying = false;
-
-    //jQuery elements
-    this.fixtures = $('#fixtures');
-    this.structure = $('#structure');
-    this.section = this.structure.find('section');
-    this.topscorers = $('#topscorers');
 }
 
+/**
+ * Stop following a team
+ * @param {number} teamId - Id of the team to stop following
+ */
 ReplayEngine.prototype.removeFollow = function(teamId) {
     $('.' + util.teamClass(teamId)).removeClass(config.class.FOLLOWED);
     util.removeVal(settings.FOLLOW_TEAMS, parseInt(teamId));
 };
 
+/**
+ * Start following a team
+ * @param {number} teamId - Id of the team to follow
+ */
 ReplayEngine.prototype.addFollow = function(teamId) {
     $('.' + util.teamClass(teamId)).addClass(config.class.FOLLOWED);
     settings.FOLLOW_TEAMS.push(parseInt(teamId));
 };
 
+/**
+ * Determines if the match should be followed
+ * @param {Match} match - Match to check
+ * @returns {boolean}
+ */
 ReplayEngine.prototype.willFollow = function(match) {
     return (this.simulateTo === -1
             && (util.findOne([match.data.homeTeam.dbid, match.data.awayTeam.dbid], settings.FOLLOW_TEAMS)
@@ -117,19 +160,21 @@ ReplayEngine.prototype.willFollow = function(match) {
  * @param {Match} match
  */
 ReplayEngine.prototype.processMatch = function(match) {
-    if(util.matchId(match.data.dbid) === this.simulateTo) {
+    this.currentMatch = match;
+    this.$currentMatch =  $('#' + util.matchId(this.currentMatch.data.dbid));
+
+    //Stop simulating, if the desired match is reached
+    if(util.matchId(this.currentMatch.data.dbid) === this.simulateTo) {
         this.simulateTo = -1;
     }
 
-    this.currentMatch = match;
-    this.$currentMatch =  $('#' + util.matchId(this.currentMatch.data.dbid));
-    var followMatch = this.willFollow(match);
+    var followMatch = this.willFollow(this.currentMatch);
 
     if(followMatch) {
+        //Remove any match selections, when the match is being followed
+        $('.' + config.class.MATCH_SELECTED).removeClass(config.class.MATCH_SELECTED);
         this.addEvent(this.highlightMatch, followMatch);
     }
-
-    this.addEvent(this.initScore, followMatch);
 
     var goals = match.getGoals();
 
@@ -147,46 +192,36 @@ ReplayEngine.prototype.processMatch = function(match) {
 ReplayEngine.prototype.highlightMatch = function() {
     var $hometeam = $('#' + this.currentMatch.data.homeTeam.dbid);
     var $awayteam = $('#' + this.currentMatch.data.awayTeam.dbid);
+    var $homegoals = this.$currentMatch.find('.homegoals');
+    var $awaygoals = this.$currentMatch.find('.awaygoals');
 
     this.$currentMatch.addClass('highlight');
     $hometeam.addClass('highlight');
     $awayteam.addClass('highlight');
 
+    //Init the match score
+    $homegoals.html(0);
+    $awaygoals.html(0);
+
     //Scroll to the next date, if the date changes
     var date = util.formatDate(this.currentMatch.data.start);
     var $dateHeader = $('#' + date.id);
 
-    var offset_t = $dateHeader.position().top + this.fixtures.scrollTop() - $('#header').height();
+    var offset_t = $dateHeader.position().top + this.$fixtures.scrollTop() - $('#header').height();
 
-    if(!this.simulate) {
-        this.fixtures.animate({
-            scrollTop: offset_t
-        }, 2000);
-    } else {
-        this.fixtures.scrollTop(offset_t);
-    }
+    this.$fixtures.animate({
+        scrollTop: offset_t
+    }, settings.SCROLL_TO_MATCH_DURATION);
 
     this.nextEvent();
 };
 
-/**
- * EVENT
- * Inits a match score by setting it to 0-0
- */
-ReplayEngine.prototype.initScore = function() {
-    var $homegoals = this.$currentMatch.find('.homegoals');
-    var $awaygoals = this.$currentMatch.find('.awaygoals');
-
-    $homegoals.html(0);
-    $awaygoals.html(0);
-
-    this.nextEvent();
-};
 
 /**
  * EVENT
  * Adds a goal and plays the replay, if it exists
  * @param event - Goal event
+ * @param {boolean} followMatch - Whether the match is followed
  * @returns {Function}
  */
 ReplayEngine.prototype.goal = function(event, followMatch) {
@@ -197,9 +232,6 @@ ReplayEngine.prototype.goal = function(event, followMatch) {
     return function() {
         var $homegoals = $match.find('.homegoals');
         var $awaygoals = $match.find('.awaygoals');
-
-
-
 
         if(followMatch) {
             //Handle goal highlights
@@ -246,7 +278,6 @@ ReplayEngine.prototype.goal = function(event, followMatch) {
                     })
                     .on('timeupdate', function() {
                        if(!rendered && this.currentTime / this.duration > 0.5) {
-                           console.log("halfway");
                            render();
                            rendered = true;
                        }
@@ -303,14 +334,11 @@ ReplayEngine.prototype.endMatch = function() {
 
     //0-0 draws are not processed by the goal event. Take care of them here.
     if(this.currentMatch.data.awayGoals + this.currentMatch.data.homeGoals === 0) {
-        this.competition.update(this.currentMatch, this.simulate);
+        this.competition.update(this.currentMatch, this.willFollow(this.currentMatch));
     }
 
     this.nextEvent();
 };
-
-
-
 
 
 /**
@@ -344,11 +372,12 @@ ReplayEngine.prototype.resume = function() {
 /**
  * Adds an event to the event queue
  * @param event
+ * @param {boolean} followMatch - Whether the match is followed
  */
 ReplayEngine.prototype.addEvent = function(event, followMatch) {
     this.events.push({
         fn: event,
-        delay: followMatch ? 2000: 0
+        delay: followMatch ? settings.EVENT_DELAY: 0
     });
 };